+++ controllers/ProductController.php
@@ -9,6 +9,7 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
+use app\models\User;
 
 /**
  * ProductController implements the CRUD actions for Product model.
@@ -70,6 +71,9 @@ class ProductController extends Controller
      */
     public function actionCreate()
     {
+        if(Yii::$app->user->identity->group > User::GROUP_KL){
+            throw new NotFoundHttpException('The requested page does not exist.');
+        }
         $model = new Product();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -101,6 +105,9 @@ class ProductController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->identity->group > User::GROUP_KL){
+            throw new NotFoundHttpException('The requested page does not exist.');
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -120,6 +127,9 @@ class ProductController extends Controller
      */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->identity->group > User::GROUP_KL){
+            throw new NotFoundHttpException('The requested page does not exist.');
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
